+++ node/clients/orders.ts
@@ -1,0 +1,20 @@
+import { ExternalClient, InstanceOptions, IOContext } from '@vtex/api'
+
+const FOUR_SECONDS = 4 * 1000
+
+export default class OrdersClient extends ExternalClient {
+  constructor(ctx: IOContext, options?: InstanceOptions) {
+    super(`http://${ctx.account}.vtexcommercestable.com.br`, ctx, {
+      ...options,
+      timeout: FOUR_SECONDS,
+    })
+  }
+
+  public getOrder = (orderId: string): Promise<Order> =>
+    this.http.get(`/api/oms/pvt/admin/orders/${orderId}`, {
+      headers: {
+        VtexIdclientAutCookie: this.context.authToken,
+      },
+      metric: 'ordersClient-getOrder',
+    })
+}