+++ node/typings/global.d.ts
@@ -1,0 +1,16 @@
+interface Settings {
+  fullCatalogImport: boolean
+  listrakToken: ListrakToken
+  listrakId?: string
+  listrakSecret?: string
+}
+
+interface DefaultSettings extends Settings {
+  listrakToken: ListrakToken | null
+}
+
+interface ListrakToken {
+  access_token: string
+  token_type: string
+  expires_in: number
+}