+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Send Sku/Order/Product data to Listrak API
+
 ## [1.3.0] - 2021-03-18
 
 ### Added